backend/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asgiref==3.11.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.3
cryptography==45.0.5
cycler==0.12.1
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
effdet==0.4.1
faiss-cpu==1.12.0
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.60.1
fsspec==2025.7.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.1
idna==3.10
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.9
layoutparser==0.3.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
omegaconf==2.3.0
openai==1.100.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
portalocker==3.2.0
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.5
pypdfium2==4.30.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pyttsx3==2.99
pytz==2025.2
pywin32==311
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.9
requests==2.32.4
rich==14.2.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.18
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.0
typer==0.20.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
wheel==0.45.1
